# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My TravelBlog API",
        "description": "API documentation for the TravelBlogJP project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "description": "Retrieve a list of all categories with their article counts",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Categories successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "description": "Create a new category with a name, slug, and optional parent category",
                "operationId": "createCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/categories/{categoryId}/edit": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Show edit form for a category",
                "description": "Display the form to edit an existing category",
                "operationId": "editCategory",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category edit form successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/categories/{categoryId}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update an existing category",
                "description": "Update the details of an existing category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "description": "Delete an existing category from the system",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category successfully deleted"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/favorite-lists": {
            "post": {
                "tags": [
                    "Favorite Lists"
                ],
                "summary": "Create a new favorite list",
                "description": "Create a new favorite list for the authenticated user.",
                "operationId": "storeFavoriteList",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Favorite list created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FavoriteList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/favorite-lists/{favoriteListId}": {
            "put": {
                "tags": [
                    "Favorite Lists"
                ],
                "summary": "Update an existing favorite list",
                "description": "Update the name of a specific favorite list.",
                "operationId": "updateFavoriteList",
                "parameters": [
                    {
                        "name": "favoriteListId",
                        "in": "path",
                        "description": "ID of the favorite list to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Favorite list updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FavoriteList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "403": {
                        "description": "Forbidden, unauthorized action"
                    },
                    "404": {
                        "description": "Favorite list not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Favorite Lists"
                ],
                "summary": "Delete a favorite list",
                "description": "Delete a specific favorite list.",
                "operationId": "deleteFavoriteList",
                "parameters": [
                    {
                        "name": "favoriteListId",
                        "in": "path",
                        "description": "ID of the favorite list to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite list deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden, cannot delete default list"
                    },
                    "404": {
                        "description": "Favorite list not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Category",
                "description": "文章分類資料結構，支持階層分類",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "旅遊"
                    },
                    "slug": {
                        "type": "string",
                        "example": "travel"
                    },
                    "parent_id": {
                        "description": "父分類ID，null代表無父分類",
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-10T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "FavoriteList": {
                "title": "FavoriteList",
                "description": "使用者收藏清單資料結構",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "收藏清單所屬使用者ID",
                        "type": "integer",
                        "example": 123
                    },
                    "name": {
                        "type": "string",
                        "example": "旅遊收藏"
                    },
                    "description": {
                        "type": "string",
                        "example": "收集所有旅遊相關文章",
                        "nullable": true
                    },
                    "articles_count": {
                        "description": "收藏清單中文章數量",
                        "type": "integer",
                        "example": 15
                    },
                    "is_default": {
                        "description": "是否為使用者的預設收藏清單",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-01T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-10T10:30:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}